fix: remove file if exists

+++ credentials.rb
@@ -32,6 +32,6 @@ module Credentials
 
   def self.delete_credentials_file_path
     credentials_file_path = "#{Dir.home}/.gem/credentials"
-    FileUtils.rm(credentials_file_path)
+    FileUtils.rm(credentials_file_path) if File.exist?(credentials_file_path)
   end
 end
